Moved showError to Utils
